added comments to all my classes in util

+++ src/core/ConfigValueContainer.cc
@@ -44,7 +44,7 @@
 
 namespace orxonox
 {
-    const int MAX_VECTOR_INDEX  = 255; // to avoid up to 4*10^9 vector entries in the config file after accidentally using a wrong argument
+    const unsigned int MAX_VECTOR_INDEX  = 255; // to avoid up to 4*10^9 vector entries in the config file after accidentally using a wrong argument
 
     /**
         @brief Initializes the ConfigValueContainer with defaultvalues.
@@ -303,7 +303,7 @@ namespace orxonox
         if (token.size() > 0)
             success = ConvertValue(&index, token[0]);
 
-        if (!success || index < 0 || index > MAX_VECTOR_INDEX)
+        if (!success || index < 0 || index > (signed int)MAX_VECTOR_INDEX)
         {
             if (!success)
             {
